COMP2411/Project/OSS Project/test/instructionTest/AdministratorTest/AdministratorSearchProductOnIDExecutorTest.java
package instructionTest.AdministratorTest;

import model.instruction.administrator.AdministratorSearchProductOnIDExecutor;
import org.junit.Before;
import org.junit.Test;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.InputStream;
import java.io.PrintStream;
import java.sql.SQLException;

import static org.junit.Assert.assertEquals;

public class AdministratorSearchProductOnIDExecutorTest {

    private AdministratorSearchProductOnIDExecutor executor;
    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();


    @Before
    public void setUpStreams() {
        System.setOut(new PrintStream(outContent));
    }

    @Before
    public void setUp() {
        executor = new AdministratorSearchProductOnIDExecutor("123"); // Use a valid phone number for testing
    }

    @Test
    public void testInformation() throws SQLException {
        executor.information();
        assertEquals("", outContent.toString());
    }

    @Test
    public void testChoose() throws SQLException {
        executor.choose("input");
        assertEquals("", outContent.toString());
    }

    @Test
    public void testGetProductID() throws SQLException {
        setInputScanner("1234567890");
        String productID = executor.getProductID();
        assertEquals("1234567890", productID);
    }

    private void setInputScanner(String input) {
        System.setIn(new ByteArrayInputStream(input.getBytes()));
    }
}

